Move section edit logic to its own component
